requirements.txt
Flask==1.1.1
requests==2.22.0
numpy==1.20.2
pandas==1.2.4
pandasql==0.7.3
xlrd==1.2.0
libopencv==0.0.1
opencv-python-headless==4.5.2.54
Pillow==8.2.0
matplotlib==3.4.2
